Refactor Dashboard components (#135)
* Refactor AllocationBlock, move to separate components and add hooks

* Refactor AllTimeBlock and CollectionsCardList
Move to separate folder, improve readable and create custom hooks

* Prettier formatting

* Refactor AvailableToBorrow component

* Refactor MyLoans components

* Refactor DashboardBorrowTab, create CardList component

* Rename "NO_DATA_CHART" => "NO_DATA_CHART_DATA"

* Rewrite useAllTimeBlock to function

* Remove unnecessary min width on singleBarChart

* Rename "CardList" to "CardsList"
Refactor conditional rendering in CardsList and AvailableToBorrow

* Add useMemo for optimization in NFTCard component

* Delete components file in DashboardBorrowTab

* Take stats to useMyLoans
