+++ src/pages/DashboardPage/components/DashboardBorrowTab/components/AvailableToBorrow/AvailableToBorrow.module.less
@@ -1,0 +1,59 @@
+.availableToBorrow {
+  display: flex;
+  flex-direction: column;
+  flex-grow: 1;
+
+  @media (max-width: 640px) {
+    margin-bottom: 16px;
+  }
+}
+
+.userNFTsValue {
+  display: flex;
+  align-items: center;
+  gap: 4px;
+
+  color: var(--content-secondary);
+  font: var(--important-text-lg);
+
+  span {
+    font: var(--table-header-md);
+  }
+
+  @media (max-width: 640px) {
+    font: var(--important-text-md);
+  }
+}
+
+.stats {
+  display: flex;
+  align-items: center;
+  flex-grow: 1;
+  gap: 16px;
+  min-height: 148px;
+  margin: 12px 0 16px;
+  width: 100%;
+
+  & > div {
+    flex-direction: column;
+    justify-content: space-between;
+    align-items: center;
+    height: 64px;
+    width: 100%;
+  }
+
+  @media (max-width: 640px) {
+    margin: 8px 0 16px;
+    min-height: unset;
+
+    & > div {
+      height: unset;
+    }
+  }
+}
+
+.button,
+.navigateLink {
+  text-decoration: none;
+  width: 100%;
+}
