+++ src/pages/DashboardPage/components/DashboardBorrowTab/components/AvailableToBorrow/index.ts
@@ -1,0 +1,1 @@
+export { default } from './AvailableToBorrow'
