+++ src/pages/DashboardPage/components/DashboardBorrowTab/components/MyLoans/MyLoans.module.less
@@ -1,0 +1,78 @@
+.loansContainer {
+  display: flex;
+  flex-direction: column;
+  flex-grow: 1;
+}
+
+.loansContent {
+  display: flex;
+  align-items: flex-end;
+  justify-content: space-between;
+  margin-bottom: 16px;
+  gap: 24px;
+  min-height: 136px;
+  height: 100%;
+  width: 100%;
+}
+
+.loansStatsContainer {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  gap: 8px;
+  height: 100%;
+  width: 100%;
+}
+
+.loansStats {
+  display: flex;
+  align-items: center;
+  gap: 16px;
+
+  @media (max-width: 640px) {
+    margin-top: 8px;
+  }
+}
+
+.loansChartStats {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+
+  & > div:nth-child(odd) {
+    background: var(--bg-primary-deep);
+  }
+}
+
+.desktopChart {
+  @media (max-width: 640px) {
+    display: none;
+  }
+}
+
+.mobileChart {
+  display: none;
+
+  @media (max-width: 640px) {
+    display: flex;
+    justify-content: center;
+  }
+}
+
+.manageLoansButton {
+  margin-bottom: 24px;
+  width: 100%;
+
+  @media (max-width: 640px) {
+    margin-bottom: 16px;
+  }
+}
+
+.doughnutChart {
+  width: 136px;
+  height: 136px;
+}
+
+.highlightLiquidation {
+  background: var(--additional-red-secondary) !important;
+}
