+++ src/pages/DashboardPage/components/DashboardBorrowTab/components/MyLoans/constants.ts
@@ -4,19 +4,19 @@ export enum LoansStatus {
   Liquidation = 'liquidation',
 }
 
-export const LOANS_COLOR_MAP: Record<LoansStatus, string> = {
+export const STATUS_COLOR_MAP: Record<LoansStatus, string> = {
   [LoansStatus.Active]: 'var(--additional-green-primary)',
   [LoansStatus.Terminating]: 'var(--additional-lava-primary)',
   [LoansStatus.Liquidation]: 'var(--additional-red-primary)',
 }
 
-export const LOANS_DISPLAY_NAMES: Record<LoansStatus, string> = {
+export const STATUS_DISPLAY_NAMES: Record<LoansStatus, string> = {
   [LoansStatus.Active]: 'Active Loans',
   [LoansStatus.Terminating]: 'Terminating Loans',
   [LoansStatus.Liquidation]: 'Liquidation during 24h',
 }
 
-export const EMPTY_DOUGHNUT_CHART_DATA = {
+export const NO_DATA_CHART_DATA = {
   value: [100],
   colors: ['var(--bg-secondary)'],
 }
