+++ src/pages/DashboardPage/components/DashboardBorrowTab/components/MyLoans/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MyLoans'
