+++ src/pages/DashboardPage/components/DashboardLendTab/DashboardLendTab.module.less
@@ -1,7 +1,3 @@
-.collectionsSection.fullWidth {
-  width: 100%;
-}
-
 .collectionsSection,
 .additionalContentSection {
   display: flex;
@@ -13,6 +9,10 @@
     flex-grow: unset;
     width: 100%;
   }
+}
+
+.collectionsSection.fullWidth {
+  width: 100%;
 }
 
 .collectionsCardList {
@@ -40,121 +40,5 @@
     & > div {
       min-width: 154px;
     }
-  }
-}
-.allTimeContainer,
-.allocationContainer {
-  display: flex;
-  flex-direction: column;
-  flex-grow: 1;
-  height: 100%;
-}
-
-.allTimeContainer {
-  @media (max-width: 640px) {
-    margin-top: 16px;
-  }
-}
-.allTimeContent {
-  display: flex;
-  align-items: flex-end;
-  margin-bottom: 24px;
-  gap: 16px;
-  width: 100%;
-  height: 100%;
-}
-.allTimeStatsContainer {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  gap: 16px;
-  min-height: 188px;
-  height: 100%;
-  width: 100%;
-
-  @media (max-width: 640px) {
-    margin-top: 8px;
-  }
-}
-.allTimeStats {
-  display: flex;
-  align-items: center;
-  gap: 16px;
-}
-.allTimeChartStats {
-  display: flex;
-  flex-direction: column;
-
-  & > div:nth-child(odd) {
-    background: var(--bg-primary-deep);
-  }
-}
-
-.allocationContent {
-  display: flex;
-  align-items: flex-end;
-  justify-content: space-between;
-  flex-grow: 1;
-  gap: 24px;
-}
-.allocationStatsContainer {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  min-height: 160px;
-  height: 100%;
-  width: 100%;
-
-  @media (max-width: 640px) {
-    margin-top: 8px;
-  }
-}
-.allocationStats {
-  display: flex;
-  align-items: center;
-  gap: 16px;
-}
-.allocationChartStats {
-  display: flex;
-  flex-direction: column;
-
-  & > div:nth-child(odd) {
-    background: var(--bg-primary-deep);
-  }
-}
-
-.doughnutChart {
-  width: 144px;
-  height: 144px;
-}
-
-.manageOffersButton {
-  margin-top: 16px;
-  width: 100%;
-}
-
-.chartContainer {
-  @media (max-width: 640px) {
-    display: none;
-  }
-}
-
-.mobileChartContainer {
-  display: none;
-
-  @media (max-width: 640px) {
-    display: flex;
-    justify-content: center;
-  }
-}
-
-.singleBarChart {
-  width: 144px;
-  min-height: 188px;
-  max-height: 188px;
-  height: 100%;
-
-  @media (max-width: 640px) {
-    width: 48px;
   }
 }
