+++ src/pages/DashboardPage/components/DashboardLendTab/components/AllocationBlock/constants.ts
@@ -1,0 +1,22 @@
+export enum AllocationStatus {
+  ActiveLoans = 'activeLoans',
+  UnderWaterLoans = 'underWaterLoans',
+  PendingOffers = 'pendingOffers',
+}
+
+export const STATUS_COLOR_MAP: Record<AllocationStatus, string> = {
+  [AllocationStatus.ActiveLoans]: 'var(--additional-green-primary)',
+  [AllocationStatus.UnderWaterLoans]: 'var(--additional-lava-primary)',
+  [AllocationStatus.PendingOffers]: 'var(--additional-blue-primary)',
+}
+
+export const STATUS_DISPLAY_NAMES: Record<AllocationStatus, string> = {
+  [AllocationStatus.ActiveLoans]: 'Active Loans',
+  [AllocationStatus.UnderWaterLoans]: 'Underwater Loans',
+  [AllocationStatus.PendingOffers]: 'Pending Offers',
+}
+
+export const NO_DATA_CHART_DATA = {
+  value: [100],
+  colors: ['var(--bg-secondary)'],
+}
