+++ src/pages/DashboardPage/components/DashboardLendTab/components/AllocationBlock/index.ts
@@ -1,0 +1,1 @@
+export { default } from './AllocationBlock'
