+++ src/pages/DashboardPage/components/DashboardLendTab/components/CollectionsCardList.tsx
@@ -1,0 +1,47 @@
+import { FC } from 'react'
+
+import { useNavigate } from 'react-router-dom'
+
+import { MarketPreview } from '@banx/api/core'
+import { PATHS } from '@banx/router'
+import { useMarketsURLControl } from '@banx/store'
+
+import { LendCard } from '../../Card'
+
+import styles from '../DashboardLendTab.module.less'
+
+interface CollectionsCardListProps {
+  marketsPreview: MarketPreview[]
+}
+
+const CollectionsCardList: FC<CollectionsCardListProps> = ({ marketsPreview }) => {
+  const { setSelectedMarkets, setMarketVisibility } = useMarketsURLControl()
+
+  const navigate = useNavigate()
+
+  const goToSelectedMarket = (collectionName: string) => {
+    setMarketVisibility(collectionName, true)
+    setSelectedMarkets([collectionName])
+
+    return navigate({
+      pathname: PATHS.LEND,
+      search: `?opened=${collectionName}&collections=${collectionName}`,
+    })
+  }
+
+  return (
+    <div className={styles.collectionsCardList}>
+      {marketsPreview.map((market) => (
+        <LendCard
+          key={market.marketPubkey}
+          image={market.collectionImage}
+          amountOfLoans={market.activeBondsAmount}
+          offerTvl={market.offerTvl}
+          apr={market.marketApr}
+          onClick={() => goToSelectedMarket(market.collectionName)}
+        />
+      ))}
+    </div>
+  )
+}
+export default CollectionsCardList
